Seperated UserComponent from BusinessComponent
