+++ src/main/java/com/example/demo/BusinessComponent/Domain/Services/BusinessService.java
@@ -14,8 +14,8 @@ import com.example.demo.BusinessComponent.API.DTOs.UpdateBusinessDTO;
 import com.example.demo.BusinessComponent.Domain.Entities.Business;
 import com.example.demo.BusinessComponent.Helpers.Mappers.BusinessMapper;
 import com.example.demo.BusinessComponent.Repositories.IBusinessRepository;
-import com.example.demo.BusinessComponent.Repositories.IUserRepository;
 import com.example.demo.CommonDTOs.PageinationDTO;
+import com.example.demo.UserComponent.Repositories.IUserRepository;
 
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
@@ -24,7 +24,7 @@ import lombok.AllArgsConstructor;
 
 @Service
 @AllArgsConstructor
-//TODO: Add exception handling
+// TODO: Add exception handling
 public class BusinessService {
   private final IBusinessRepository businessRepository;
   private final IUserRepository userRepository;
@@ -32,12 +32,13 @@ public class BusinessService {
 
   public BusinessDTO createBusiness(@NotNull @Valid CreateBusinessDTO dto) {
     Business business = businessMapper.toBusiness(dto);
-    Business savedBusiness = businessRepository.save(business); 
+    Business savedBusiness = businessRepository.save(business);
     return businessMapper.toBusinessDTO(savedBusiness);
   }
 
   public GetBusinessListDTO getBusinessList(@NotNull @Valid PageinationDTO pageinationInfo) {
-    //Set up pageable entity that is used for retrieving the specific business items from the database
+    // Set up pageable entity that is used for retrieving the specific business
+    // items from the database
     Pageable pageable = PageRequest.of(pageinationInfo.getPage(), pageinationInfo.getPageSize());
     Page<Business> pageination = businessRepository.findAll(pageable);
     return businessMapper.toGetBusinessListDTO(pageination);
@@ -49,15 +50,15 @@ public class BusinessService {
   }
 
   public BusinessDTO updateBusiness(@NotNull UUID businessId, @NotNull @Valid UpdateBusinessDTO updateBusinessDTO) {
-    //TODO: Move exception handling to a different class
+    // TODO: Move exception handling to a different class
     Business existingBusiness = businessRepository.findById(businessId)
-            .orElseThrow(() -> new EntityNotFoundException("Business not found with ID: " + businessId));
+        .orElseThrow(() -> new EntityNotFoundException("Business not found with ID: " + businessId));
 
     if (updateBusinessDTO.getOwnerId() != null) {
-        boolean ownerExists = userRepository.existsById(updateBusinessDTO.getOwnerId());
-        if (!ownerExists) {
-            throw new EntityNotFoundException("No user found with ID: " + updateBusinessDTO.getOwnerId());
-        }
+      boolean ownerExists = userRepository.existsById(updateBusinessDTO.getOwnerId());
+      if (!ownerExists) {
+        throw new EntityNotFoundException("No user found with ID: " + updateBusinessDTO.getOwnerId());
+      }
     }
 
     Business updatedBusiness = businessMapper.updateBusinessFromDto(updateBusinessDTO, existingBusiness);
@@ -66,7 +67,7 @@ public class BusinessService {
   }
 
   public void deleteBusiness(@NotNull UUID businessId) {
-    //TODO: Move exception handling to a different class
+    // TODO: Move exception handling to a different class
     if (businessRepository.existsById(businessId)) {
       businessRepository.deleteById(businessId);
     } else {
