+++ src/main/java/com/example/demo/UserComponent/API/DTOs/CreateUserDTO.java
@@ -1,12 +1,13 @@
-package com.example.demo.BusinessComponent.API.DTOs;
+package com.example.demo.UserComponent.API.DTOs;
 
 import java.util.UUID;
 
-import com.example.demo.BusinessComponent.Domain.Entities.Enums.RoleType;
+import com.example.demo.UserComponent.Domain.Entities.Enums.RoleType;
 
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -15,18 +16,22 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor
-public class UpdateUserDTO {
-  
+public class CreateUserDTO {
+
+  @NotNull
   private final String fullName;
 
+  @NotNull
   @Pattern(regexp = "\\+\\d{3,30}")
   private final String phoneNumber;
 
+  @NotNull
   @Email
   private final String emailAddress;
 
   private final UUID businessId;
 
   @Enumerated(EnumType.STRING)
+  @NotNull
   private final RoleType role;
 }
