+++ src/main/java/com/example/demo/UserComponent/API/Endpoints/UserController.java
@@ -1,12 +1,12 @@
-package com.example.demo.BusinessComponent.API.Endpoints;
+package com.example.demo.UserComponent.API.Endpoints;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.demo.BusinessComponent.API.DTOs.CreateUserDTO;
-import com.example.demo.BusinessComponent.API.DTOs.UpdateUserDTO;
-import com.example.demo.BusinessComponent.API.DTOs.UserDTO;
-import com.example.demo.BusinessComponent.ApplicationServices.BusinessApplicationService;
+import com.example.demo.UserComponent.API.DTOs.CreateUserDTO;
+import com.example.demo.UserComponent.API.DTOs.UpdateUserDTO;
+import com.example.demo.UserComponent.API.DTOs.UserDTO;
+import com.example.demo.UserComponent.ApplicationServices.UserApplicationService;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
@@ -24,37 +24,34 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-
-
-
 @RestController
 @AllArgsConstructor
 @RequestMapping("/v1/user")
 public class UserController {
-  private final BusinessApplicationService businessApplicationService;
+  private final UserApplicationService userApplicationService;
 
   @PostMapping
   public ResponseEntity<UserDTO> createUser(@Valid @RequestBody CreateUserDTO createUserDTO) {
-    UserDTO createdUser = businessApplicationService.createUser(createUserDTO);
+    UserDTO createdUser = userApplicationService.createUser(createUserDTO);
     return ResponseEntity.status(HttpStatus.CREATED).body(createdUser);
   }
-  
+
   @GetMapping("/{userId}")
-  public ResponseEntity<UserDTO> getMethodName(@PathVariable UUID userId) {
-    UserDTO gottenUserDTO = businessApplicationService.getUser(userId);
+  public ResponseEntity<UserDTO> getUser(@PathVariable UUID userId) {
+    UserDTO gottenUserDTO = userApplicationService.getUser(userId);
     return ResponseEntity.ok().body(gottenUserDTO);
   }
 
   @PutMapping("/{userId}")
-  public ResponseEntity<UserDTO> putMethodName(@PathVariable UUID userId, @RequestBody UpdateUserDTO updateUserDTO) {
-    UserDTO updatedUserDTO = businessApplicationService.updateUser(userId, updateUserDTO);
+  public ResponseEntity<UserDTO> updateUser(@PathVariable UUID userId, @RequestBody UpdateUserDTO updateUserDTO) {
+    UserDTO updatedUserDTO = userApplicationService.updateUser(userId, updateUserDTO);
     return ResponseEntity.ok().body(updatedUserDTO);
   }
 
   @DeleteMapping("/{userId}")
   public ResponseEntity<Void> deleteUser(@NotNull @PathVariable UUID userId) {
-    businessApplicationService.deleteUser(userId);
+    userApplicationService.deleteUser(userId);
     return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
   }
-  
+
 }
