+++ src/main/java/com/example/demo/UserComponent/Domain/Services/UserService.java
@@ -1,29 +1,27 @@
-package com.example.demo.BusinessComponent.Domain.Services;
+package com.example.demo.UserComponent.Domain.Services;
 
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
-import com.example.demo.BusinessComponent.Repositories.IUserRepository;
+import com.example.demo.UserComponent.API.DTOs.CreateUserDTO;
+import com.example.demo.UserComponent.API.DTOs.UpdateUserDTO;
+import com.example.demo.UserComponent.API.DTOs.UserDTO;
+import com.example.demo.UserComponent.Domain.Entities.User;
+import com.example.demo.UserComponent.Helpers.Mappers.UserMapper;
+import com.example.demo.UserComponent.Repositories.IUserRepository;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
-
-import com.example.demo.BusinessComponent.API.DTOs.CreateUserDTO;
-import com.example.demo.BusinessComponent.API.DTOs.UpdateUserDTO;
-import com.example.demo.BusinessComponent.API.DTOs.UserDTO;
-import com.example.demo.BusinessComponent.Domain.Entities.User;
-import com.example.demo.BusinessComponent.Helpers.Mappers.UserMapper;
-
 import lombok.AllArgsConstructor;
 
 @Service
 @AllArgsConstructor
-//TODO: Add exception handling
+// TODO: Add exception handling
 public class UserService {
   private final IUserRepository userRepository;
   private final UserMapper userMapper;
-  
+
   public UserDTO createUser(@NotNull @Valid CreateUserDTO createUserDTO) {
     User user = userMapper.toUser(createUserDTO);
     User savedUser = userRepository.save(user);
@@ -42,7 +40,7 @@ public class UserService {
   }
 
   public void deleteUser(@NotNull UUID userId) {
-    //TODO:Add checking if user id null
+    // TODO:Add checking if user id null
     userRepository.deleteById(userId);
   }
 }
